requirements.txt
# Core
openai>=1.0.0
python-dotenv>=1.0.0
jsonschema>=4.0.0
pymongo>=4.0.0
loguru>=0.7.0
python-dateutil>=2.8.0
typing-extensions>=4.0.0

# Google Services
gspread>=5.12.0
oauth2client>=4.1.3
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1

# Data Processing
pandas>=2.0.0
openpyxl>=3.1.0
pdf2image>=1.16.0
pytesseract>=0.3.10
Pillow>=10.0.0

# Automation & Scheduling
schedule>=1.2.0
watchdog>=3.0.0
requests>=2.31.0
click>=8.1.0

# PDF Generation
reportlab>=4.0.0

# Security & Secrets
keyring>=24.0.0

# Optional / Advanced (Uncomment if needed)
# numpy>=1.24.0
# scikit-learn>=1.3.0
